Enable font-lock for correct indentation
The Font Lock mode is involved in erlang-mode's indentation rules and
is normally enabled by default for all modes. But when emacs is running
without a display, as in batch mode, Font Lock mode is never enabled.

By calling `font-lock-fontify-buffer` we will get correct indentation
for the parts of the code handled by font-lock.

Signed-off-by: Björn Svensson <bjorn.a.svensson@est.tech>